+++ mycpp/marksweep_heap.h
@@ -132,8 +132,8 @@ class MarkSweepHeap {
     root_set_.RootInCurrentFrame(root);
   }
 
-  void* Allocate(int num_bytes);
-  void* Reallocate(void* p, int num_bytes);
+  void* Allocate(size_t num_bytes);
+  void* Reallocate(void* p, size_t num_bytes);
   int Collect();
   void MarkObjects(Obj* obj);
   void Sweep();
